#177 Adds Button Variant to LinkTo
